CLI tests: Automatically skip unsupported features.
Use the test262 features field to check whether the testcase claims to
use any unsupported features. Skip as appropriate.

The set of features that the interpreter supports are stored in
test/helpers/supported-features.js. The set of all features that test262
use can be generated by the generate-all-features.js helper.