Reverse again iteration order for `tuple-binders`
The previous fix was correct but would leave the list of binders in
reverse order (but with correctly assigned index). This broke pseudo
generation that only looks at the binders and not the assigned index.

This generates a list in the same order than the initial list, but the
the correct indices.

Fixes #24